create job small refactor; wallet action fix
