+++ lib/create_job/expanded_selection.dart
@@ -1,0 +1,61 @@
+import 'package:flutter/services.dart';
+import 'package:provider/provider.dart';
+import '../blockchain/interface.dart';
+import 'package:flutter/material.dart';
+
+class ExpandedSection extends StatefulWidget {
+  final VoidCallback callback;
+
+  final Widget child;
+  final bool expand;
+  const ExpandedSection({super.key, this.expand = false, required this.child, required this.callback});
+
+  @override
+  _ExpandedSectionState createState() => _ExpandedSectionState();
+}
+
+class _ExpandedSectionState extends State<ExpandedSection> with SingleTickerProviderStateMixin {
+  late AnimationController expandController;
+  late Animation<double> animation;
+
+  @override
+  void initState() {
+    super.initState();
+    prepareAnimations();
+    _runExpandCheck();
+  }
+
+  void prepareAnimations() {
+    expandController = AnimationController(vsync: this, duration: const Duration(milliseconds: 200));
+    animation = CurvedAnimation(
+      parent: expandController,
+      curve: Curves.fastOutSlowIn,
+    );
+  }
+
+  void _runExpandCheck() {
+    if (widget.expand) {
+      expandController.forward();
+    } else {
+      expandController.reverse();
+    }
+  }
+
+  @override
+  void didUpdateWidget(ExpandedSection oldWidget) {
+    super.didUpdateWidget(oldWidget);
+    _runExpandCheck();
+  }
+
+  @override
+  void dispose() {
+    expandController.dispose();
+    super.dispose();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    widget.callback();
+    return SizeTransition(axisAlignment: 1.0, sizeFactor: animation, child: widget.child);
+  }
+}
