+++ lib/navigation/beamtest.dart
@@ -1,19 +1,19 @@
-import 'package:beamer/beamer.dart';
-import 'package:flutter/src/widgets/framework.dart';
-
-class HomeLocation extends BeamLocation {
-  // @override
-  // List<BeamPage> get pages => [CustomerPageWidget.beamLocation];
-  //
-  // @override
-  // String get pathBlueprint => CustomerPageWidget.path;
-
-  @override
-  List<BeamPage> buildPages(
-      BuildContext context, RouteInformationSerializable state) {
-    throw UnimplementedError();
-  }
-
-  @override
-  List<Pattern> get pathPatterns => throw UnimplementedError();
-}
+// import 'package:beamer/beamer.dart';
+// import 'package:flutter/src/widgets/framework.dart';
+//
+// class HomeLocation extends BeamLocation {
+//   // @override
+//   // List<BeamPage> get pages => [CustomerPageWidget.beamLocation];
+//   //
+//   // @override
+//   // String get pathBlueprint => CustomerPageWidget.path;
+//
+//   @override
+//   List<BeamPage> buildPages(
+//       BuildContext context, RouteInformationSerializable state) {
+//     throw UnimplementedError();
+//   }
+//
+//   @override
+//   List<Pattern> get pathPatterns => throw UnimplementedError();
+// }
